# AgriBridge-Connect/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriBridge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  farmer: "farmer",
  customer: "customer",
  storage_manager: "storage_manager",
} as const;

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
  phone?: string;
  role: RegisterRequestRole;
  farmName?: string;
  location?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  phone?: string;
  role: string;
  location?: string;
  farmName?: string;
  bio?: string;
  profileImage?: string;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UpdateProfileRequest {
  name?: string;
  phone?: string;
  location?: string;
  farmName?: string;
  bio?: string;
  profileImage?: string;
}

export interface UpdateFcmTokenRequest {
  fcmToken: string;
}

export interface Product {
  id: number;
  farmerId: number;
  name: string;
  description?: string;
  price: string;
  quantity: string;
  unit: string;
  category: string;
  imageUrl?: string;
  location?: string;
  available: boolean;
  createdAt?: string;
  updatedAt?: string;
}

export interface ProductWithFarmer {
  id: number;
  farmerId: number;
  name: string;
  description?: string;
  price: string;
  quantity: string;
  unit: string;
  category: string;
  imageUrl?: string;
  location?: string;
  available: boolean;
  createdAt?: string;
  updatedAt?: string;
  farmerName?: string;
  farmName?: string;
  farmerPhone?: string;
  averageRating?: number;
  reviewCount?: number;
}

export interface AddProductRequest {
  name: string;
  description?: string;
  price: number;
  quantity: number;
  unit?: string;
  category: string;
  imageUrl?: string;
  location?: string;
}

export interface UpdateProductRequest {
  name?: string;
  description?: string;
  price?: number;
  quantity?: number;
  unit?: string;
  category?: string;
  imageUrl?: string;
  location?: string;
  available?: boolean;
}

export interface Order {
  id: number;
  customerId: number;
  farmerId: number;
  productId: number;
  quantity: string;
  totalAmount: string;
  status: string;
  deliveryAddress?: string;
  notes?: string;
  paymentId?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface OrderWithDetails {
  id: number;
  customerId: number;
  farmerId: number;
  productId: number;
  quantity: string;
  totalAmount: string;
  status: string;
  deliveryAddress?: string;
  notes?: string;
  paymentId?: string;
  createdAt?: string;
  updatedAt?: string;
  productName?: string;
  productImage?: string;
  productCategory?: string;
  farmerName?: string;
  customerName?: string;
}

export interface PlaceOrderRequest {
  productId: number;
  quantity: number;
  deliveryAddress?: string;
  notes?: string;
  paymentId?: string;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  packed: "packed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export interface ColdStorage {
  id: number;
  managerId: number;
  name: string;
  description?: string;
  location: string;
  totalCapacity: string;
  availableCapacity: string;
  pricePerKgPerDay: string;
  temperature?: string;
  phone?: string;
  imageUrl?: string;
  available: boolean;
  createdAt?: string;
}

export interface ColdStorageWithManager {
  id: number;
  managerId: number;
  name: string;
  description?: string;
  location: string;
  totalCapacity: string;
  availableCapacity: string;
  pricePerKgPerDay: string;
  temperature?: string;
  phone?: string;
  imageUrl?: string;
  available: boolean;
  createdAt?: string;
  managerName?: string;
}

export interface CreateStorageRequest {
  name: string;
  description?: string;
  location: string;
  totalCapacity: number;
  pricePerKgPerDay: number;
  temperature?: string;
  phone?: string;
  imageUrl?: string;
}

export interface Booking {
  id: number;
  farmerId: number;
  storageId: number;
  managerId: number;
  productName: string;
  quantity: string;
  startDate: string;
  endDate: string;
  totalCost?: string;
  status: string;
  notes?: string;
  createdAt?: string;
}

export interface BookingWithDetails {
  id: number;
  farmerId: number;
  storageId: number;
  managerId: number;
  productName: string;
  quantity: string;
  startDate: string;
  endDate: string;
  totalCost?: string;
  status: string;
  notes?: string;
  createdAt?: string;
  storageName?: string;
  storageLocation?: string;
  farmerName?: string;
}

export interface BookStorageRequest {
  storageId: number;
  productName: string;
  quantity: number;
  startDate: string;
  endDate: string;
  notes?: string;
}

export type ApproveBookingRequestStatus =
  (typeof ApproveBookingRequestStatus)[keyof typeof ApproveBookingRequestStatus];

export const ApproveBookingRequestStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface ApproveBookingRequest {
  status: ApproveBookingRequestStatus;
}

export interface Review {
  id: number;
  customerId: number;
  productId: number;
  orderId?: number;
  rating: number;
  comment?: string;
  createdAt?: string;
}

export interface ReviewWithCustomer {
  id: number;
  customerId: number;
  productId: number;
  orderId?: number;
  rating: number;
  comment?: string;
  createdAt?: string;
  customerName?: string;
}

export interface AddReviewRequest {
  productId: number;
  orderId?: number;
  rating: number;
  comment?: string;
}

export interface FarmerStats {
  totalProducts: number;
  totalOrders: number;
  totalEarnings: number;
  storageBookings: number;
  pendingOrders: number;
}

export type GetProductsParams = {
  category?: string;
  minPrice?: number;
  maxPrice?: number;
  location?: string;
  farmerId?: number;
};