# AgriBridge-Connect/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgriBridge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all available products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: location
          in: query
          schema:
            type: string
        - name: farmerId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductWithFarmer"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductWithFarmer"

    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /orders/place:
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderRequest"
      responses:
        "200":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [farmer, customer, storage_manager]
        farmName:
          type: string
        location:
          type: string
      required:
        - email
        - password
        - name
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
        location:
          type: string
        farmName:
          type: string
        bio:
          type: string
        profileImage:
          type: string
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
...
    Product:
      type: object
      properties:
        id:
          type: integer
        farmerId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: string
        quantity:
          type: string
        unit:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        location:
          type: string
        available:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - farmerId
        - name
        - price
        - quantity
        - unit
        - category
        - available
    ProductWithFarmer:
      type: object
      properties:
        id:
          type: integer
        farmerId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: string
        quantity:
          type: string
        unit:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        location:
          type: string
        available:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        farmerName:
          type: string
        farmName:
          type: string
        farmerPhone:
          type: string
        averageRating:
          type: number
        reviewCount:
          type: integer
      required:
        - id
        - farmerId
        - name
        - price
        - quantity
        - unit
        - category
        - available
...
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: number
        unit:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        location:
          type: string
        available:
          type: boolean
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        farmerId:
          type: integer
        productId:
          type: integer
        quantity:
          type: string
        totalAmount:
          type: string
        status:
          type: string
        deliveryAddress:
          type: string
        notes:
          type: string
        paymentId:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerId
        - farmerId
        - productId
        - quantity
        - totalAmount
        - status
...
    PlaceOrderRequest:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: number
        deliveryAddress:
          type: string
        notes:
          type: string
        paymentId:
          type: string
      required:
        - productId
        - quantity
    UpdateOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, packed, shipped, delivered, cancelled]
      required:
        - status
...