# AgriBridge-Connect/lib/api-zod/src/generated/types/updateProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriBridge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProductRequest {
  name?: string;
  description?: string;
  price?: number;
  quantity?: number;
  unit?: string;
  category?: string;
  imageUrl?: string;
  location?: string;
  available?: boolean;
}