# AgriBridge-Connect/lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AgriBridge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  phone?: string;
  role: string;
  location?: string;
  farmName?: string;
  bio?: string;
  profileImage?: string;
  createdAt?: string;
}